pyproject.toml
[project]
name = "rsscord"
version = "0.1.0"
description = "Simple Discord RSS bot with SQLite history"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.5.0,<3.0.0",
    "feedparser>=6.0.11,<7.0.0",
]

[tool.uv]
package = false
